# expression: handle max_allowed_packet warnings for space function.