## What have you changed? (mandatory)
Return NULL and a warning when the result of space exceeds max_allowed_packetbs.

Before this PR:
```
tidb> select space(18446744073709551617);
+-----------------------------+
| space(18446744073709551617) |
+-----------------------------+
| NULL                        |
+-----------------------------+
1 row in set, 1 warning (0.00 sec)

tidb> show warnings;
+---------+------+-----------------------------------------------------------+
| Level   | Code | Message                                                   |
+---------+------+-----------------------------------------------------------+
| Warning | 1292 | Truncated incorrect DECIMAL value: '18446744073709551617' |
+---------+------+-----------------------------------------------------------+
1 row in set (0.00 sec)
```
After this PR:
```
tidb> select space(18446744073709551617);
+-----------------------------+
| space(18446744073709551617) |
+-----------------------------+
| NULL                        |
+-----------------------------+
1 row in set, 2 warnings (0.00 sec)

tidb> show warnings;
+---------+------+-----------------------------------------------------------------------------+
| Level   | Code | Message                                                                     |
+---------+------+-----------------------------------------------------------------------------+
| Warning | 1292 | Truncated incorrect DECIMAL value: '18446744073709551617'                   |
| Warning | 1301 | Result of space() was larger than max_allowed_packet (67108864) - truncated |
+---------+------+-----------------------------------------------------------------------------+
2 rows in set (0.00 sec)
```

## What is the type of the changes? (mandatory)
Bug fix (non-breaking change which fixes an issue)

## How has this PR been tested? (mandatory)
- unit test
- explain test

## Does this PR affect documentation (docs/docs-cn) update? (mandatory)
no

## Does this PR affect tidb-ansible update? (mandatory)
no

## Does this PR need to be added to the release notes? (mandatory)
Yes

release note:
```
Return `NULL` when the result of function `SPACE` is larger than `max_allowed_packet`
```


## Refer to a related PR or issue link (optional)
to [#7153](https://github.com/pingcap/tidb/issues/7153)
## Benchmark result if necessary (optional)

## Add a few positive/negative examples (optional)